extract branch count
