+++ frontend/app/src/entities/branches/api/get-branches-count-from-api.ts
@@ -1,7 +1,7 @@
 import graphqlClient from "@/shared/api/graphql/graphqlClientApollo";
 import type { PaginationParams } from "@/shared/api/types";
 
-import { GET_BRANCHES } from "@/entities/branches/api/get-branches-query";
+import { GET_BRANCHES_COUNT } from "@/entities/branches/api/get-branches-count-query";
 
 export const BRANCHES_PER_PAGE = 40;
 
@@ -9,17 +9,11 @@ export interface GetBranchesFromApiParams extends PaginationParams {
   branchName?: string;
 }
 
-export const getBranchesFromApi = async ({
-  branchName,
-  limit = BRANCHES_PER_PAGE,
-  offset,
-}: GetBranchesFromApiParams = {}) => {
+export const getBranchesCountFromApi = async (branchName?: string) => {
   return graphqlClient.query({
-    query: GET_BRANCHES,
+    query: GET_BRANCHES_COUNT,
     variables: {
       branchName,
-      limit,
-      offset,
     },
   });
 };
