+++ frontend/app/src/entities/branches/api/get-branches-count-query.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const GET_BRANCHES_COUNT = gql`
+query GetBranchesCount($branchName: String) {
+    InfrahubBranch(name__value: $branchName) {
+      count
+    }
+  }
+`;
