+++ frontend/app/src/entities/branches/domain/get-branches.ts
@@ -3,7 +3,6 @@ import { store } from "@/shared/stores";
 
 import {
   type GetBranchesFromApiParams,
-  getBranchesCountFromApi,
   getBranchesFromApi,
 } from "@/entities/branches/api/get-branches-from-api";
 import {
@@ -48,16 +47,4 @@ export const getAllBranches = async (): Promise<GetBranchesResult> => {
   store.set(branchesState, branches);
 
   return branches;
-};
-
-// Fetch branches count (for header badge)
-export const getBranchesCount = async (branchName?: string): Promise<number> => {
-  const { data, errors } = await getBranchesCountFromApi(branchName);
-
-  if (errors) {
-    throw new Error(errors.map((e) => e.message).join("; "));
-  }
-
-  const response = data as InfrahubBranchResponse;
-  return response?.InfrahubBranch?.count ?? 0;
 };
